Readmes & helper scripts

+++ install_deps_centos.sh
@@ -2,47 +2,61 @@
 
 set -e
 
-sudo yum update
-sudo yum install git cmake mc tmux
-sudo yum install autoconf automake libtool curl make unzip wget
-sudo yum install clang
+mkdir -p lib
+
+sudo yum update -y
+sudo yum install -y \
+    git cmake mc tmux \
+    autoconf automake libtool curl make unzip wget \
+    clang gcc-c++ -y
 
 export CC=/usr/bin/gcc
 export CXX=/usr/bin/g++
 
-cd ../intel-cmt-cat/lib/
+pushd `pwd`
+cd third_party/intel-cmt-cat/lib/
 make SHARED=n
 cp libpqos.a ../../../lib/
+popd
 
-export CC=/usr/bin/clang
-export CXX=/usr/bin/clang++
-
-#curl $curlopts -O https://googlemock.googlecode.com/files/gmock-1.7.0.zip
-cd third_party/protobuf
+if [ ! -d "./third_party/protobuf" ]
+then
+pushd `pwd`
+cd ./third_party/protobuf
 wget http://pkgs.fedoraproject.org/repo/pkgs/gmock/gmock-1.7.0.zip/073b984d8798ea1594f5e44d85b20d66/gmock-1.7.0.zip
 unzip -q gmock-1.7.0.zip
 rm gmock-1.7.0.zip
 mv gmock-1.7.0 gmock
+popd
+fi
 
+pushd `pwd`
+cd ./third_party/protobuf
 ./autogen.sh
 ./configure
 make -j2
-#make check
 sudo make install
-sudo ldconfig # refresh shared library cache
+popd
 
-cd ../grpc
-cd grpc
+pushd `pwd`
+cd ./third_party/grpc
 git submodule update --init
-make
+make -j2
 sudo make install
+popd
 
-cd ../snap-plugin-lib-cpp
-mkdir build
+export CC=/usr/bin/clang
+export CXX=/usr/bin/clang++
+
+pushd `pwd`
+cd ./third_party/snap-plugin-lib-cpp
+mkdir -p build
 ./autogen.sh
 ./configure -prefix=`pwd`/build
-make
+make -j2
 sudo make install
 cp build/lib/libsnap.a ../../lib
+popd
 
-export LD_LIBRARY_PATH=/usr/local/lib
+sudo cp /usr/local/lib/* /usr/lib
+sudo ldconfig -v
